Allow deploying only HTML changes
Now if no packages are listed to be built, it will check for possible
changes in the generated HTML files. This allows, in tandem with being
able to skip come CI steps when the change is only in the deploy script,
to skip rebuilding all packages, which wastes lots of resources and most
of the time isn't necessary.